move mobile sidebar button into header
